common: Break out filesystem helpers into their own component
