+++ common/find_runfiles.cc
@@ -1,11 +1,8 @@
 #include "drake/common/find_runfiles.h"
 
-#include <sys/stat.h>
-#include <unistd.h>
-
 #include <cstdlib>
-#include <fstream>
 #include <memory>
+#include <stdexcept>
 
 #ifdef __APPLE__
 #include <mach-o/dyld.h>
@@ -15,6 +12,7 @@
 #include "tools/cpp/runfiles/runfiles.h"
 
 #include "drake/common/drake_assert.h"
+#include "drake/common/filesystem.h"
 #include "drake/common/never_destroyed.h"
 #include "drake/common/text_logging.h"
 
@@ -192,36 +190,6 @@ RlocationOrError FindRunfile(const std::string& resource_path) {
       "Sought '{}' in runfiles directory '{}' {}; "
       "perhaps a 'data = []' dependency is missing.",
       resource_path, singleton.runfiles_dir, detail);
-  return result;
-}
-
-bool IsFile(const std::string& filesystem_path) {
-  struct stat attributes{};
-  int error = stat(filesystem_path.c_str(), &attributes);
-  if (error) { return false; }
-  return S_ISREG(attributes.st_mode);
-}
-
-bool IsDir(const std::string& filesystem_path) {
-  struct stat attributes{};
-  int error = stat(filesystem_path.c_str(), &attributes);
-  if (error) { return false; }
-  return S_ISDIR(attributes.st_mode);
-}
-
-std::string Readlink(const std::string& pathname) {
-  std::string result;
-  result.resize(4096);
-  ssize_t length = ::readlink(pathname.c_str(), &result.front(), result.size());
-  if (length < 0) {
-    throw std::runtime_error(fmt::format(
-        "Could not open {}", pathname));
-  }
-  if (length >= static_cast<ssize_t>(result.size())) {
-    throw std::runtime_error(
-        fmt::format("Could not readlink {} (too long)", pathname));
-  }
-  result.resize(length);
   return result;
 }
